Extract LobbyService from the lobby websocket API
Lobby state, business rules, publishing, and the countdown/game-load
machinery move into a DI-registered LobbyService (the chat-service
pattern: @singleton, TypedPublisher, CodedError subclass). The wsapi
LobbyApi becomes a transport shell: it validates nydus payloads,
resolves the acting user/client sockets, delegates to the service, and
converts LobbyServiceError codes back to the exact http-errors (status,
message, and LobbyCreateErrorCode/LobbyJoinErrorCode bodies) it threw
before, so the wire behavior is unchanged.

Service methods take plain validated params rather than nydus data maps,
so an HTTP API can call them directly when lobby mutations move off
websocket RPC. Methods whose bodies are fully synchronous are typed
sync, which also keeps their errors catchable at the call site.

Join gets its own AlreadyStarted/AlreadyInActivity error codes because
the wire behavior genuinely differs per operation: join failures carry
machine-readable body codes that create's equivalents never had.
